Handle ID return values. Fixes #191
