+++ lib/composite_primary_keys.rb
@@ -48,6 +48,7 @@ require 'active_record/associations/preloader/association'
 require 'active_record/associations/preloader/belongs_to'
 require 'active_record/associations/preloader/has_and_belongs_to_many'
 
+
 require 'active_model/dirty'
 
 require 'active_record/attribute_methods/dirty'
@@ -67,6 +68,7 @@ require 'active_record/validations/uniqueness'
 
 # CPK files
 require 'composite_primary_keys/persistence'
+require 'composite_primary_keys/active_record_overides'
 require 'composite_primary_keys/base'
 require 'composite_primary_keys/core'
 require 'composite_primary_keys/composite_arrays'
